requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
wheel
py-cord
discord-py-slash-command
scipy
sympy
youtube-dl
